% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pileup.R
\name{annotate_pileup}
\alias{annotate_pileup}
\title{Annotating pileup}
\usage{
annotate_pileup(pileup, exon, input.type = "whole_intron",
  output.type = "part_intron", intron.len = NULL)
}
\arguments{
\item{pileup}{raw coverage data}

\item{exon}{exon}

\item{input.type}{type of intronic region contained in pileup,
with choices "whole_intron", "part_intron", or "only_exon"; the first is the default.}

\item{output.type}{type of intronic region that will be included in output,
with choices "whole_intron", "part_intron", or "only_exon"; the second is the default.}

\item{intron.len}{a length of intronic region that will be included in output.}
}
\description{
This function is used to get pileup data based on new transcript
for the purpose of the downstream outlier detection analysis.
}
