% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoosePCs.R
\name{ChoosePCs}
\alias{ChoosePCs}
\title{A function to automatically choose the number of principal components.}
\usage{
ChoosePCs(seurat.obj = NULL, cutoff = NULL)
}
\arguments{
\item{cutoff}{The cutoff value for cumulative proportion of variance explained. Can be set by the user, or can be determine automatically. Defaults to NULL.}

\item{seurat.object}{The object containing our single cell counts and principal component matrix. Defaults to NULL.}
}
\description{
This function uses the eigenvalues of the principal component matrix to determine the best number of PCs. The default can be chosen automatically, or given by the user. It is intended to be run after \code{RunPCA()}.
}
\examples{
ChoosePCs(seurat.obj = pbmc, cutoff = .75)
}
