% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertGeneOrthologs.R
\name{ConvertGeneOrthologs}
\alias{ConvertGeneOrthologs}
\title{Convert a vector of MGI symbols to their HGNC orthologs and vice versa.}
\usage{
ConvertGeneOrthologs(gene.vec = NULL, species = "mm")
}
\arguments{
\item{gene.vec}{A vector of genes to convert. Default to NULL.}

\item{species}{One of "mm" or "hs". Defaults to "mm" (and thus converts to human).}
}
\description{
Converts gene names from human -> mouse and mouse -> human.
}
\examples{
\dontrun{ConvertGeneOrthologs(gene.vec = mouse_genes)}
\dontrun{ConvertGeneOrthologs(gene.vec = human_genes, species = "hs")}
}
\author{
Jack Leary
}
