% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineDist.R
\name{CosineDist}
\alias{CosineDist}
\title{A function to compute the cosine distance matrix.}
\usage{
CosineDist(input = NULL)
}
\arguments{
\item{input}{The input matrix. Defaults to NULL.}
}
\description{
This function takes a matrix as input, and computes the cosine distance (1 - cosine similarity) between the observations.
}
\examples{
CosineDist(input = pca_matrix)
}
