% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosSimilarity.R
\name{CosSimilarity}
\alias{CosSimilarity}
\title{A function to compute the cosine similarity of two vectors.}
\usage{
CosSimilarity(x = NULL, y = NULL)
}
\arguments{
\item{x}{Input vector. Default is NULL.}

\item{y}{Input vector. Default is NULL}
}
\description{
This function takes as input two vectors of length \emph{n}, \emph{x} and \emph{y}, for which the cosine similarity is computed.
}
\examples{
CosSimilarity(x, y)
}
