% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindFilteredMarkers.R
\name{FindFilteredMarkers}
\alias{FindFilteredMarkers}
\title{This function finds marker genes for a given set of cells, and filters out highly expressed genes from cells in a reference Seurat object.}
\usage{
FindFilteredMarkers(
  obj.1 = NULL,
  obj.2 = NULL,
  ident.1 = "seurat_clusters",
  ident.2 = "seurat_clusters",
  test.use = "wilcox",
  pval.cutoff = 0.05,
  logfc.cutoff = 0.25,
  quantile.cutoff = 0.9,
  extra.cell.filter = NULL
)
}
\description{
Used to filter candidate marker genes for a subcluster Seurat object. Prevents markers from being chosen if they are highly expressed in other celltypes.
}
\examples{
\dontrun{filtered_markers <- FindFilteredMarkers(obj.1 = subclust_obj, obj.2 = full_obj, ident.1 = "label", ident.2 = "seurat_clusters")}
}
\seealso{
\code{\link{FindSpecificMarkers}}

\code{\link[Seurat]{FindAllMarkers}}
}
\author{
Jack Leary
}
