% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSpecificMarkers.R
\name{FindSpecificMarkers}
\alias{FindSpecificMarkers}
\title{This functions finds specific marker genes for a all clusters in a \code{Seurat} object.}
\usage{
FindSpecificMarkers(
  seurat.object = NULL,
  ident.use = "seurat_clusters",
  de.method = "wilcox",
  perc.cutoff = 0.9,
  log2fc.cutoff = 0.25,
  fdr.cutoff = 0.05
)
}
\arguments{
\item{seurat.object}{The \code{Seurat} object containing clusters for which you'd like marker genes identified. Defaults to NULL.}

\item{ident.use}{The cell identity to group by. Defaults to "seurat_clusters".}

\item{de.method}{The differential expression method used in \code{FindAllMarkers()}. Defaults to "wilcox".}

\item{perc.cutoff}{The percentile cutoff used to find highly expressed genes in other cluster. Defaults to 0.9.}

\item{log2fc.cutoff}{The log2FC cutoff used, in part, to determine whether a gene is differentially expressed. Defaults to 0.25.}

\item{fdr.cutoff}{The cutoff used to remove DE genes with non-significant adjusted \emph{p}-values. Defaults to 0.05.}
}
\description{
This function finds marker genes for all clusters, and then filters those markers on a per-cluster basis against the most highly expressed genes in other clusters.
}
\examples{
FindSpecificMarkers(seurat_object, method = "wilcox")
}
