% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindSubpopulationMarkers.R
\name{FindSubpopulationMarkers}
\alias{FindSubpopulationMarkers}
\title{Identify marker genes for previosuly identified subpopulations.}
\usage{
FindSubpopulationMarkers(
  seurat.object = NULL,
  reclust.data = NULL,
  which.clust = NULL,
  diff.exp.test = "wilcox",
  logfc.thresh = 2,
  random.seed = 629
)
}
\arguments{
\item{seurat.object}{The original \code{Seurat} object containing the entire cell population and related metadata.}

\item{reclust.data}{The list of reclustered \code{Seurat} objects as returned by \code{ReclusterCells}.}

\item{which.clust}{The cluster (with subpopulations) for which you'd like to obtain marker genes with respect to the entire cell population.}

\item{diff.exp.test}{The test used to calculate differential expression using \code{FindMarkers}. Defaults to "wilcox".}

\item{logfc.thresh}{The lo2 fold-change cutoff used when performing differential expression analysis. Defaults to 2.}

\item{random.seed}{(Optional) The seed used to control stochasticity in several functions. Defaults to 629.}
}
\description{
This function determines which cells characterize the subpopulations identified using \code{ReclusterCells}. It is intended to be run on a single re-clustered \code{Seurat} object at a time, though if you wish you could
iterate over the list of reclustering results, and save the outputs from this function in a matching array of lists. The function returns a list of dataframes, one dataframe per cluster, containing normal and Bonferroni-adjust
p-values, gene prevalence, and effect size in the form of log2 fold change.
}
\examples{
FindSubpopulationMarkers(seurat.object = pbmc_small, reclust.data = reclust_results)
FindSubpopulationMarkers(seurat.object = pbmc_small, reclust.data = reclust_results, diff.exp.test = "t")
}
