% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offstat.R
\name{get_offstat}
\alias{get_offstat}
\title{Step 2 outlier detection: get_offstat}
\usage{
get_offstat(resmat, rawmat, exonset, winlength = 100, readconstr = 10)
}
\description{
Step 2 outlier detection: get_offstat
}
