% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateSubclusters.R
\name{IntegrateSubclusters}
\alias{IntegrateSubclusters}
\title{A function to integrate subcluster identities into the original Seurat object.}
\usage{
IntegrateSubclusters(
  original.object = NULL,
  reclust.results = NULL,
  do.plot = FALSE
)
}
\arguments{
\item{original.object}{The original Seurat object. Defaults to NULL.}

\item{reclust.results}{A list of reclustering results as output from ReclusterCells(). Defaults to NULL.}

\item{do.plot}{Should the results be plotted on a dimension reduction plot? Defaults to FALSE.}
}
\description{
This function takes a list of outputs from ReclusterCells() and integrates the new subcluster identities into the original Seurat object.
}
\examples{
IntegrateSubclusters(original.object = pbmc, reclust.results = my_subclusts)
}
\author{
Jack Leary
}
