% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtransform_data.R
\name{logtransform_data}
\alias{logtransform_data}
\title{Log-transforming data}
\usage{
logtransform_data(data, logshift.val = NULL, ...)
}
\arguments{
\item{data}{data to be transformed}
}
\description{
This function is used to get
}
\keyword{log}
\keyword{parameter}
\keyword{shift}
