% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscGlobal.R
\name{miscGlobal}
\alias{miscGlobal}
\title{Detect global shape variants in RNA-seq data}
\usage{
miscGlobal(data, siglev = 1e-10, subt.mean = FALSE, PCnum = NULL,
  maxPCnum = 20, eps = NULL, rm.PCdir = TRUE, ADcutoff = 10,
  filter.dir = TRUE, less.return = TRUE)
}
\arguments{
\item{data}{precessed RNA-seq data from \link{process_data}}

\item{siglev}{the significance level of the Chi-squared distribution. Default is 1e-10.}

\item{subt.mean}{logical, whether to subtract mean before SVD. Default is FALSE.}

\item{PCnum}{the number of PCs to be used.
If NULL (the default) the number of PCs will be estimated.}

\item{maxPCnum}{the maximum number of PCs to be used. Default is 20.}

\item{eps}{tuning parameter for estimating the number of PCs.
If NULL (default), eps = (1/nrow(data))}

\item{rm.PCdir}{logical or numeric, the indices of PCs to be excluded from further study.
If TRUE (the default), this automatically excludes a set of PCs that deviate from normality.
If a sequence is specified, the corresponding PCs will be excluded.}

\item{ADcutoff}{a cutoff value for checking the normality based on Anderson-Darling test statistic}

\item{filter.dir}{logical, wether to filter out directions that deviate from normality.
Default is TRUE.}

\item{less.return}{logical, whether to show less results and not to return large matrices.
Default is TRUE.}
}
\description{
This function discovers outlying subjects whose RNA-seq have abnormal shapes
and provides the most outlying direction for each outlier.
}
