% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscLocal.R
\name{miscLocal}
\alias{miscLocal}
\title{Step 2 outlier detection}
\usage{
miscLocal(miscGlobalobj, covermat, exonset, winlength = 100,
  readconstr = 10, siglev = 1e-05, cutoff = NULL)
}
\description{
Step 2 outlier detection
}
