% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceDimensions.R
\name{ReduceDimensions}
\alias{ReduceDimensions}
\title{A function to run PCA / t-SNE / UMAP / PHATE.}
\usage{
ReduceDimensions(obj = NULL, n.PC = NULL, which.algs = NULL, seed = NULL)
}
\arguments{
\item{obj}{The Seurat object to run dimension reduction algorithms on. Defaults to NULL.}

\item{n.PC}{How many principal components should be used? Can be an integer or "auto". Defaults to NULL.}

\item{which.algs}{Which nonlinear dimension algorithms can be used? Should be some combination of "tsne", "umap", and "phate". Defaults to NULL.}

\item{seed}{The random seed to use. Defaults to NULL.}
}
\description{
This function simplifies the running of various dimension reduction algorithms. It exists mostly to make the body of ReclusterCells() easier to read.
}
\examples{
ReduceDimensions(pbmc3k, n.PC = 10, which.algs = "umap", seed = 629)
}
