% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSingleR.R
\name{RunSingleR}
\alias{RunSingleR}
\title{A function to run \code{SingleR} on a \code{Seurat} object.}
\usage{
RunSingleR(seurat.object = NULL, ref.data = NULL, dif.exp.method = "wilcox")
}
\arguments{
\item{seurat.object}{The \code{Seurat} object containing the cells you'd like to ID.}

\item{ref.data}{(Optional) The reference dataset whose labels you'd like to assign to your cells. Defaults to the Human Primary Cell Atlas.}

\item{dif.exp.method}{The differential expression testing method used by \code{SingleR}. Defaults to the Wilcoxon ranked-sum test.}
}
\description{
This function runs the \code{SingleR} cell type identification algorithm, with a \code{Seurat} object as input.
}
\examples{
callSingleR(seurat.object = pbmc)
callSingleR(seurat.object = pbmc, ref.data = my_ref, dif.exp.method = "t")
}
\references{
Aran \emph{et al} (2019). Reference-based analysis of lung single-cell sequencing data reveals a transitional profibrotic macrophage.
}
