% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.factor.R
\name{scale.factor}
\alias{scale.factor}
\title{Obtain median scale factors based on a adjusted center}
\usage{
\method{scale}{factor}(X, average = "mean", trim = 0.1, adjval = NULL)
}
\description{
Obtain median scale factors based on a adjusted center
}
