% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_yehlab.R
\name{theme_yehlab}
\alias{theme_yehlab}
\title{A clean \code{ggplot2} theme for dimension reduction plots.}
\usage{
theme_yehlab()
}
\value{
A function specifying a minimalist \code{ggplot2} theme.
}
\description{
This is the Yeh Lab's default \code{ggplot2} theme for dimension reduction scatterplots, and was used throughout the SCISSORS manuscript. Like all \code{ggplot2}-based themes, you can add more themes to it, or override them if you wish.
}
\examples{
\dontrun{
DimPlot(pbmc) + theme_yehlab()
}
}
\seealso{
\code{\link[ggplot2]{theme}}
}
\author{
Jack Leary
}
