% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSilhouetteScores.R
\name{ComputeSilhouetteScores}
\alias{ComputeSilhouetteScores}
\title{Calculate the mean silhouette score of a clustering.}
\usage{
ComputeSilhouetteScores(seurat.obj = NULL)
}
\arguments{
\item{seurat.obj}{The input object for which silhouette score will be computed. Defaults to NULL.}
}
\description{
This function will compute the silhouette score for each cluster identified by \code{Seurat}'s Louvain modularity optimization community detection algorithm.
}
\examples{
ComputeSilhouetteScores(seurat.obj)
}
