% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CosineDist.R
\name{CosineDist}
\alias{CosineDist}
\title{Compute the cosine distance matrix.}
\usage{
CosineDist(input.mat = NULL)
}
\arguments{
\item{input.mat}{The input matrix. Defaults to NULL.}
}
\description{
This function takes a matrix as input, and computes the cosine distance (1 - cosine similarity) between the observations.
}
\examples{
\dontrun{
CosineDist(input.mat = pca_matrix)
}
}
\author{
Jack Leary
}
