% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_onoffout.R
\name{detect_onoffout}
\alias{detect_onoffout}
\title{Step 2 outlier detection: combine two types of outliers}
\usage{
detect_onoffout(resmat, rawmat, exonset, winlength = 100, readconstr = 10,
  siglev = NULL, cutoff = NULL)
}
\description{
Step 2 outlier detection: combine two types of outliers
}
