% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateSubclusters.R
\name{IntegrateSubclusters}
\alias{IntegrateSubclusters}
\title{Integrate estimated subcluster identities into the original Seurat object.}
\usage{
IntegrateSubclusters(
  original.object = NULL,
  reclust.results = NULL,
  do.plot = FALSE
)
}
\arguments{
\item{original.object}{The original Seurat object. Defaults to NULL.}

\item{reclust.results}{Either an individual \code{Seurat} object or a list of reclustering results as output from \code{\link{ReclusterCells}}. Defaults to NULL.}

\item{do.plot}{Should the results be plotted on a dimension reduction plot? Defaults to FALSE.}
}
\value{
A \code{Seurat} object.
}
\description{
This function takes a list of outputs from \code{\link{ReclusterCells}} and integrates the new subcluster identities into the original Seurat object.
}
\examples{
\dontrun{
IntegrateSubclusters(original.object = pbmc, reclust.results = my_subclusts)
}
}
\seealso{
\code{\link{ReclusterCells}}
}
\author{
Jack Leary
}
