% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare scRNA-seq data for reclustering.}
\usage{
PrepareData(
  seurat.object = NULL,
  use.sct = TRUE,
  n.HVG = 4000,
  regress.mt = TRUE,
  regress.cc = TRUE,
  n.PC = "auto",
  var.cutoff = 0.15,
  which.dim.reduc = c("umap"),
  perplexity = 30,
  umap.lr = 0.05,
  initial.resolution = 0.3,
  nn.metric = "cosine",
  k.val = NULL,
  do.plot = NULL,
  random.seed = 629
)
}
\arguments{
\item{seurat.object}{The object containing the cells you'd like to analyze. Defaults to NULL.}

\item{use.sct}{Should \code{SCTransform} be used for normalization / HVG selection? Defaults to TRUE, otherwise typical log-normalization is used.}

\item{n.HVG}{The number of highly variable genes to compute. Defaults to 4000.}

\item{regress.mt}{Should the percentage of mitochondrial DNA be computed and regressed out? Works for mouse / human gene names. Defaults to TRUE.}

\item{regress.cc}{Should cell cycle scores be computed & regressed out? NOTE: uses human cell cycle genes. Defaults to TRUE.}

\item{n.PC}{The number of PCs used as input to non-linear dimension reduction and clustering algorithms. Can be chosen by user, or set automatically using \code{\link{ChoosePCs}}. Defaults to "auto".}

\item{var.cutoff}{(Optional) The proportion of variance explained cutoff to be used when n.PC is set to "auto". Defaults to .15.}

\item{which.dim.reduc}{(Optional) Which non-linear dimension reduction algorithms should be used? Supports "tsne", "umap", "phate", and "all". Plots will be generated using the t-SNE embedding. Defaults to c("umap"), as most users will likely not have \code{phateR} installed.}

\item{perplexity}{(Optional) What perplexity value should be used when embedding cells in t-SNE space? Defaults to 30.}

\item{umap.lr}{(Optional) What learning rate should be used for the UMAP embedding? Defaults to 0.05.}

\item{initial.resolution}{The initial resolution parameter used in the \code{\link[Seurat]{FindAllMarkers}} function. Defaults to 0.3.}

\item{nn.metric}{(Optional) The distance metric to be used in computing the SNN graph. Defaults to "cosine".}

\item{k.val}{(Optional) The nearest-neighbors parameter \emph{k} to be used when creating the shared nearest-neighbor graph. Defaults to \eqn{k \approx \sqrt{n}}.}

\item{do.plot}{(Optional) The dimension reduction view you'd like plotted. Should be one of "tsne", "umap", "phate", or "pca". Defaults to NULL.}

\item{random.seed}{The seed used to control stochasticity in several functions. Defaults to 629.}
}
\description{
This function prepares  single cell data for reclustering analysis. The input is a \code{Seurat} object in any stage of pre-processing, or even a \code{SingleCellExperiment} object that will be converted to \code{Seurat} format. The function checks which metadata features (\% mitochondrial DNA, cell cycle scores) and assays are present (normalized counts, PCA & t-SNE embeddings), then runs an initial graph-based clustering.
}
\examples{
\dontrun{PrepareData(seurat.object, n.variable.genes = 3000, n.PC = 20, do.plot = TRUE)}
\dontrun{PrepareData(seurat.object, initial.resolution = .5, k.val = 25, random.seed = 100)}
}
\references{
Stuart \emph{et al} (2019). Comprehensive integration of single-cell data. \emph{Cell}.
}
\seealso{
\code{\link{ChoosePCs}}

\code{\link[Seurat]{FindAllMarkers}}
}
\author{
Jack Leary
}
