% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareData.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare scRNA-seq data for reclustering.}
\usage{
PrepareData(
  seurat.object = NULL,
  n.variable.genes = 4000,
  n.PC = 30,
  which.dim.reduc = c("tsne", "umap"),
  perplexity = 30,
  initial.resolution = 0.3,
  k.val = NULL,
  do.plot = FALSE,
  random.seed = 629
)
}
\arguments{
\item{seurat.object}{The object containing the cells you'd like to analyze.}

\item{n.variable.genes}{The number of variable genes to find at each step. Defaults to 4000.}

\item{n.PC}{The number of PCs used as input to non-linear dimension reduction and clustering algorithms. Defaults to 30.}

\item{which.dim.reduc}{(Optional) Which non-linear dimension reduction algorithms should be used? Supports "tsne", "umap", "phate", and "all". Plots will be generated using the t-SNE embedding. Defaults to c("tsne", "umap"), as most users will likely not have \code{phateR} installed.}

\item{perplexity}{(Optional) What perplexity value should be used when embedding cells in t-SNE space? Defaults to 30.}

\item{initial.resolution}{The initial resolution parameter used in the \code{FindClusters} function. Defaults to 0.3.}

\item{k.val}{(Optional) The parameter \emph{k} to be used when creating the shared nearest-neighbor graph. Defaults to \emph{k} ~ sqrt(\emph{n}).}

\item{do.plot}{(Optional) Should the function print a t-SNE plot of your cells to the graphics viewer? Defaults to FALSE.}

\item{random.seed}{The seed used to control stochasticity in several functions. Defaults to 629.}
}
\description{
This function prepares a \code{Seurat} object for reclustering analysis.
The input is a \code{Seurat} object in any stage of pre-processing, or even a \code{SingleCellExperiment} object that will be converted to \code{Seurat} format.
The function checks which metadata features and assays are present (\% mitochondrial DNA, normalized counts, PCA & t-SNE embeddings), then runs an initial graph-based clustering.
}
\examples{
PrepareData(seurat.object, n.variable.genes = 5000, do.plot = TRUE)
PrepareData(seurat.object, initial.resolution = .5, k.val = 25, random.seed = 100)
}
\references{
Stuart \emph{et al} (2019). Comprehensive integration of single-cell data. \emph{Cell}.
}
