% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclusterCells.R
\name{ReclusterCells}
\alias{ReclusterCells}
\title{Identify subpopulations in single cell clusters.}
\usage{
ReclusterCells(
  seurat.object = NULL,
  n.variable.genes = 4000,
  n.PC = 10,
  which.dim.reduc = c("tsne", "umap"),
  redo.embedding = TRUE,
  which.clust = NULL,
  resolution.vals = c(0.05, 0.1, 0.2, 0.35),
  k.val = NULL,
  do.plot = FALSE,
  random.seed = 629
)
}
\arguments{
\item{seurat.object}{The \code{Seurat} object containing cells and their assigned cluster IDs.}

\item{n.variable.genes}{How many variable genes should be detected in each subcluster? Defaults to 4000.}

\item{n.PC}{How many PCs should be used as input to non-linear to non-linear dimension reduction and clustering algorithms. Defaults to 10.}

\item{which.dim.reduc}{(Optional). Which non-linear dimension reduction algorithms should be used? Supports "tsne", "umap", "phate", and "all". Plots will be generated using the t-SNE embedding. Defaults to c("tsne", "umap"), as most users will likely not have \code{phateR} installed.}

\item{redo.embedding}{(Optional) Should a cluster-specific dimension reduction embeddings be generated? Sometimes subpopulations appear mixed together on the original coordinates, but separate clearly when re-embedded. Defaults to TRUE.}

\item{which.clust}{Which clusters should undergo subpopulation detection analysis? A user-provided list.}

\item{resolution.vals}{(Optional) A user-defined vector of resolution values to compare when clustering cells. Defaults to c(.05, .1, .15, .2, .35).}

\item{k.val}{(Optional) The parameter \emph{k} to be used when creating the shared nearest-neighbor graph. Defaults to \emph{k} ~ sqrt(\emph{n}).}

\item{do.plot}{(Optional) Should t-SNE plots of the various reclusterings be plotted for visual inspection by the user? Defaults to FALSE.}

\item{random.seed}{The seed used to control stochasticity in several functions. Defaults to 629.}
}
\description{
This function identifies subclusters of cell types by recalculating the \emph{n} most highly variable genes for each cluster using \code{sctransform} as implemented in \code{Seurat}. The function returns a list of \code{Seurat} objects, one for each cluster the user wants to investigate.
}
