% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclusterCells.R
\name{ReclusterCells}
\alias{ReclusterCells}
\title{Identify subpopulations in single cell clusters.}
\usage{
ReclusterCells(
  seurat.object = NULL,
  which.clust = NULL,
  auto = FALSE,
  merge.clusters = FALSE,
  use.parallel = TRUE,
  n.cores = 3,
  use.sct = FALSE,
  n.HVG = 4000,
  n.PC = "auto",
  redo.embedding = TRUE,
  resolution.vals = c(0.1, 0.2, 0.3, 0.4),
  k.vals = c(10, 25, 50),
  is.integrated = FALSE,
  integration.ident = NULL,
  cutoff.score = 0.25,
  nn.metric = "cosine",
  regress.mt = FALSE,
  regress.cc = FALSE,
  random.seed = 312
)
}
\arguments{
\item{seurat.object}{The \code{Seurat} object containing cells and their assigned cluster IDs.}

\item{which.clust}{Which clusters should undergo subpopulation detection analysis? A user-provided list or single integer. Leave NULL if setting \code{auto = TRUE}. Defaults to NULL.}

\item{auto}{Should the clusters on which to run SCISSORS be determined automatically? If so, \code{which.clust} will be chosen through silhouette score analysis. Not recommended for large datasets as the distance matrix calculation is computationally expensive. Defaults to FALSE.}

\item{merge.clusters}{If multiple clusters are specified, should the clusters be grouped as one before running SCISSORS? Defaults to FALSE.}

\item{use.parallel}{Should the \code{Seurat} data reprocessing & the main reclustering loop be parallelized? Defaults to TRUE.}

\item{n.cores}{The number of cores to be used in parallel computation is \code{use.parallel = TRUE}. Defaults to 3.}

\item{use.sct}{Should \code{SCTransform} be used for normalization / HVG selection? Defaults to FALSE, which equates to using typical log1p-normalization.}

\item{n.HVG}{How many variable genes should be detected in each subcluster? Defaults to 4000.}

\item{n.PC}{How many PCs should be used as input to non-linear to non-linear dimension reduction and clustering algorithms. Can be provided by the user, or set automatically by \code{\link{ChoosePCs}}. Defaults to "auto".}

\item{redo.embedding}{(Optional) Should a cluster-specific dimension reduction embeddings be generated? Sometimes subpopulations appear mixed together on the original coordinates, but separate clearly when re-embedded. Defaults to TRUE.}

\item{resolution.vals}{A user-defined vector of resolution values to compare when clustering cells. Defaults to c(.1, .2, .3, .4).}

\item{k.vals}{The values of the number of nearest neighbors \emph{k} to be tested. Defaults to c(10, 25, 50).}

\item{is.integrated}{Do the data come from multiple samples & need to be re-integrated? See https://github.com/satijalab/seurat/issues/1883 for discussion on this topic. Defaults to FALSE.}

\item{integration.ident}{If the data are to be re-integrated, what metadata column contains the sample identity? Defaults to NULL.}

\item{cutoff.score}{The lowest mean silhouette score accepted as evidence of subclusters. Defaults to .25, reasonable values are [.1, .3].}

\item{nn.metric}{(Optional) The distance metric to be used in computing the SNN graph. Defaults to "cosine".}

\item{regress.mt}{(Optional) Should the percentage of mitochondrial DNA be computed and regressed out? Works for mouse / human gene names. Defaults to FALSE.}

\item{regress.cc}{(Optional) Should cell cycle scores be computed & regressed out? NOTE: uses human cell cycle genes. Defaults to FALSE.}

\item{random.seed}{The seed used to control stochasticity in several functions. Defaults to 312}
}
\value{
If the number of clusters to be re-clustered is 1, returns the re-clustered \code{Seurat} object, else returns a list of re-clustered \code{Seurat} objects.
}
\description{
This function identifies subclusters of cell types by recalculating the \emph{n} most highly variable genes for each cluster using \code{\link[Seurat]{SCTransform}}. The function returns a list of \code{Seurat} objects, one for each cluster the user wants to investigate.
}
\examples{
\dontrun{
ReclusterCells(seurat.object,
               which.clust = 5,
               resolution.vals = c(.1, .2, .5),
               k.vals = c(10, 20, 30))
ReclusterCells(seurat.object,
               which.clust = c(0, 3, 5),
               merge.clusters = TRUE,
               n.HVG = 2000,
               n.PC = 20,
               cutoff.score = 0.1)
ReclusterCells(seurat.object,
               auto = TRUE,
               use.parallel = TRUE,
               n.cores = 6,
               is.integrated = TRUE,
               integration.ident = "samplename")
}
}
\seealso{
\code{\link{ComputeSilhouetteScores}}

\code{\link{IntegrateSubclusters}}
}
\author{
Jack Leary
}
