% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReclusterCells.R
\name{ReclusterCells}
\alias{ReclusterCells}
\title{A function to iteratively recalculate cluster assignments based on highly variable genes}
\usage{
ReclusterCells(
  seurat.object = NULL,
  recluster.res = 1,
  var.method = "sctransform",
  use.SingleR.labels = FALSE
)
}
\arguments{
\item{seurat.object}{The \code{Seurat} object containing cells and their assigned cluster IDs.}

\item{recluster.res}{The amount of times to recluster each cluster and subcluster. Defaults to 1.}

\item{var.method}{How should highly variable genes be identified? One of "sctransform" or "vst". Defaults to "sctransform".}

\item{use.SingleR.labels}{Should cell types from \code{SingleR} be used to recluster cells? Defaults to FALSE.}

\item{n.variable.genes}{How many variable genes should be detected in each subcluster? Defaults to 3000}
}
\description{
This function identifies subclusters of cell types by recalculating the \emph{n} most highly variable genes for each cluster.
}
\examples{
ReclusterCells(seurat.object)
ReclusterCells(seurat.object, recluster.res = 2, use.SingleR.labels = TRUE)
}
