% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceDimensions.R
\name{ReduceDimensions}
\alias{ReduceDimensions}
\title{Run PCA / t-SNE / UMAP / PHATE.}
\usage{
ReduceDimensions(
  obj = NULL,
  n.PC = "auto",
  which.algos = c("umap"),
  random.seed = 312
)
}
\arguments{
\item{obj}{The Seurat object to run dimension reduction algorithms on. Defaults to NULL.}

\item{n.PC}{How many principal components should be used? Can be an integer or "auto". Defaults to "auto".}

\item{which.algos}{Which nonlinear dimension algorithms can be used? Should be some combination of "tsne", "umap", and "phate". Defaults to "umap".}

\item{random.seed}{The random seed to use. Defaults to 312.}
}
\value{
A \code{Seurat} object.
}
\description{
This function simplifies the running of various dimension reduction algorithms. It exists mostly to make the body of \code{\link{ReclusterCells}} easier to read.
}
\examples{
\dontrun{
ReduceDimensions(pbmc,
                 n.PC = 10,
                 which.algos = c("umap", "phate"))
}
}
\seealso{
\code{\link{RunPHATE}}
}
\author{
Jack Leary
}
