% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPHATE.R
\name{RunPHATE}
\alias{RunPHATE}
\title{A function to perform nonlinear dimension reduction using PHATE.}
\usage{
RunPHATE(
  object = NULL,
  n.components = 2,
  n.PC = NULL,
  mds.method = "smacof",
  dist.metric = "cosine",
  random.seed = 629
)
}
\arguments{
\item{object}{The Seurat object you'd like to run PHATE on. Defaults to NULL.}

\item{n.components}{How dimensions should the data be reduced to? Defaults to 2.}

\item{n.PC}{How many PCs should be used as input to the algorithm? Defaults to NULL.}

\item{mds.method}{The solver used for MDS. Defaults to SMACOF, but SGD can be used to speed up the algorithm.}

\item{dist.metric}{The distance metric to use for KNN and MDS. Defaults to the cosine distance.}

\item{random.seed}{The random seed used to control stochasticity. Defaults to 629.}
}
\description{
This function wraps the PHATE dimension reduction algorithm in the typical Seurat syntax.
}
\examples{
RunPhate(object = pbmc3k, n.components = 2, n.PC = 10)
}
\author{
Jack Leary
}
